backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25

# Tavily API
tavily-python==0.3.3

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Scheduler
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
